+++ src/main/java/kr/hhplus/be/server/application/order/OrderApplicationService.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.application.order;
 
 import kr.hhplus.be.server.application.order.dto.criteria.OrderCriteria;
 import kr.hhplus.be.server.application.order.dto.result.OrderResult;
-import kr.hhplus.be.server.application.order.event.OrderEventPublisher;
+import kr.hhplus.be.server.domain.order.event.OrderEventPublisher;
 import kr.hhplus.be.server.support.aop.DistributedLock;
 import kr.hhplus.be.server.domain.order.service.OrderService;
 import kr.hhplus.be.server.domain.order.dto.info.OrderInfo;
